backend/requirements.txt
fastapi
uvicorn
google-generativeai
qdrant-client
python-dotenv
pydantic
